done add to cart

+++ server/yoga-business-management-web/src/main/java/org/example/yogabusinessmanagementweb/common/entities/Cart.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -16,7 +17,7 @@ import java.util.List;
 @FieldDefaults(level = lombok.AccessLevel.PRIVATE)
 public class Cart extends AbstractEntity<Long>{
     int totalItem;
-    Long totalPrice;
+    BigDecimal totalPrice;
 
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "user_id")
